app/src/requirements.txt
alembic==1.4.3
aniso8601==8.1.0
appdirs==1.4.4
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
flask-marshmallow==0.14.0
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
idna==2.10
importlib-metadata==3.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.9.1
marshmallow-sqlalchemy==0.24.1
pipenv==2020.11.15
prometheus-client==0.9.0
prometheus-flask-exporter==0.18.1
psycopg2-binary==2.8.6
pyaml==20.4.0
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
six==1.15.0
SQLAlchemy==1.3.20
urllib3==1.26.2
virtualenv==20.2.1
virtualenv-clone==0.5.4
Werkzeug==1.0.1
zipp==3.4.0
